+++ src/Providers/Default/DefaultControllerProvider.cs
@@ -1,0 +1,13 @@
+﻿using System.ComponentModel.Composition;
+
+namespace DialToolsForVS
+{
+    [Export(typeof(IDialControllerProvider))]
+    internal class DefaultControllerProvider : IDialControllerProvider
+    {
+        public IDialController TryCreateController(IDialControllerHost host)
+        {
+            return new DefaultController();
+        }
+    }
+}